# feat(python)!: ensure consistent return types from `item`