--- a/py-polars/polars/dataframe/frame.py
+++ b/py-polars/polars/dataframe/frame.py
@@ -1752,9 +1752,14 @@ class DataFrame:
             ).render()
         )
 
-    def item(self, row: int | None = None, column: int | str | None = None) -> Any:
+    def item(
+        self,
+        row: int | None = None,
+        column: int | str | None = None,
+        lists_as_series: bool = True,
+    ) -> Any:
         """
-        Return the dataframe as a scalar, or return the element at the given row/column.
+        Return the element at the given row/column, or the dataframe as a scalar.
 
         Notes
         -----
@@ -1767,6 +1772,8 @@ class DataFrame:
             Optional row index.
         column
             Optional column index or name.
+        lists_as_series
+            When retrieving nested List items, return them as Series.
 
         Examples
         --------
@@ -1789,7 +1796,7 @@ class DataFrame:
                     f"Can only call '.item()' if the dataframe is of shape (1,1), or if "
                     f"explicit row/col values are provided; frame has shape {self.shape}"
                 )
-            return self._df.select_at_idx(0).get_idx(0)
+            return self._df.select_at_idx(0).get_idx(0, lists_as_series)
 
         elif row is None or column is None:
             raise ValueError("cannot call `.item()` with only one of `row` or `column`")
@@ -1801,7 +1808,7 @@ class DataFrame:
         )
         if s is None:
             raise ValueError(f"column index {column!r} is out of bounds")
-        return s.get_idx(row)
+        return s.get_idx(row, lists_as_series)
 
     def to_arrow(self) -> pa.Table:
         """