Ensures that we return consistent types from `item`, taking the fast-path by default.

* All dtypes (scalars, array, struct etc) _except_ `List` return python-native values from `item()` - now `List` does too (instead of being the only one that returns a Series).
* This update provides a **~50%** speedup over having to make a subsequent `to_list()` call after the initial call to `item()` (as is currently necessary).

## Example
```python
import polars as pl

df = pl.DataFrame({
    "a": ["1", "2", "3"],
    "b": [["a", "b"], ["c", "d"], ["e", "f"]],
})
# shape: (3, 2)
# ┌─────┬────────────┐
# │ a   ┆ b          │
# │ --- ┆ ---        │
# │ str ┆ list[str]  │
# ╞═════╪════════════╡
# │ 1   ┆ ["a", "b"] │
# │ 2   ┆ ["c", "d"] │
# │ 3   ┆ ["e", "f"] │
# └─────┴────────────┘
```
Current behaviour:
```python
df.item(1, "b")
# shape: (2,)
# Series: '' [str]
# [
#   "c"
#   "d"
# ]
```
Updated behaviour (breaking):
```python
df.item(1, "b")
# ['c', 'd']
```

## Timings
Previously, to get python-native values back:
```python
%timeit  df.item(1, "b").to_list()
# 745 ns ± 2.69 ns per loop (mean ± std. dev. of 7 runs, 1,000,000 loops each)

%timeit  df.item(1, "b")
# 478 ns ± 2.35 ns per loop (mean ± std. dev. of 7 runs, 1,000,000 loops each)
```