@stinodego 2023-08-16
In general, I dislike extra parameters that toggle a return type. I believe we can call that an anti-pattern in Polars land. So I'm not a fan of this addition.

However, you do raise a good point that it's strange a Series is returned for List columns, when a Struct with nested lists returns a list normally:

```
>>> s = pl.Series([{'a': [1, 2], 'b': 2}])
>>> s.item()
{'a': [1, 2], 'b': 2}
```
And Arrays:
```
>>> s = pl.Series([[1,2], [3,4]], dtype=pl.Array(2, pl.Int64))
>>> s.item(0)
[1, 2]
```


If we decide `.item()` on a List column should return a regular list, I'd say we should just deprecate the old behavior and change the behavior in a breaking change.

If we decide Lists should be represented as Series, we should do so for Arrays / nested Struct types too.
And the same logic should be followed in `DataFrame.to_dict`.