@alexander-beedie 2023-08-16
> In general, I dislike extra parameters that toggle a return type. I believe we can call that an anti-pattern in Polars land. So I'm not a fan of this addition.

Indeed; but the only reason it's necessary is because we have a single outlier in terms of the return type, and I wanted to avoid making a breaking change.

However, am equally happy (actually - happier :) to avoid adding the extra param at all and just make things consistent (native python values, as returned from all the other related functions and other dtypes) and tag this as a breaking change for 0.19 👍