@alexander-beedie 2023-08-16
> @stinodego Do we just want to schedule this for 0.19? I don't know if you can make a deprecation warning for something that has no alternative because the new breaking change alternative... is the same function, just with different behavior.

Yes: _"and tag this as a breaking change for 0.19"_  ;)