@stinodego 2023-08-17
Doesn't this introduce a copy, though? Shouldn't we prefer returning `Series` everywhere in name of performance?