@alexander-beedie 2023-08-17
Closing in favour of an upcoming new issue; @stinodego and I had a good discussion and there are broader themes involved here than just `item` ;)